Remove duplicate timestamps when generate logfile -- fixes #1
Use a hash to store the number of timestamps found in the all-frames
PDB file, where its keys are the timestamps and the corresponding values
are the number of times the timestamps appear in a given project/run/
clone. This info can be used for logging/reporting purposes if needed.
However, these purposes are not implemented yet.